requirements.txt
easyocr==1.7.2
opencv-python==4.8.1.78
matplotlib==3.7.2
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.1
pandas==2.0.3
python-Levenshtein==0.21.1
scikit-image==0.21.0
Shapely==2.0.1
PyYAML==6.0.1
python-bidi==0.4.2
pyclipper==1.3.0.post6
ninja==1.11.1 